+++ modules/core.lua
@@ -2,5 +2,5 @@
 EZOArmory = EZOArmory or {}
 
 EZOArmory.ADDON_NAME    = "EZOArmory"
-EZOArmory.ADDON_VERSION = "0.1.6"
+EZOArmory.ADDON_VERSION = "0.1.7"
 EZOArmory.AUTHOR        = "@Zuriplayer"